Bandaid fix for Combat Shotgun sprite glitch
The code for shotguns references the number of shells in the sidesaddles. Since the Combat Shotgun has no sidesaddles, it tries to use the "empty" frame, which doesn't exist, so I just set it to use the default pickup sprite. This is the simplest solution I could think of, because i really don't wanna make another HDShotgun clone, lmao