Reworked reload/unload behavior
Because the Combat Shotgun is a bottom loader, there's no way to access the chamber like with the Hunter, so reloading and unloading now only affects the ammo tube. The only way to unload a chambered shell now is to rack the pump. Also, chambering a round while racked is now impossible.